se cambio a BIGINT el type de ID de los modelos Cart, Payment_method y Profile. Se agrego el modelo Users pero falta seguirlo

+++ database/models/Payment_method.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, dataTypes) => {
     let alias = "Payment_method";
     let cols = {
         id: {
-            type: dataTypes.INTEGER,
+            type: dataTypes.BIGINT,
             PrimaryKey: true,
             autoIncrement: true
         },
